+++ SymbolCollector.sln
@@ -18,6 +18,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ELFSharp", "src\ELFSharp\EL
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SymbolCollector.Android.UITests", "test\SymbolCollector.Android.UITests\SymbolCollector.Android.UITests.csproj", "{48BB7372-FB0A-46DE-B66C-FDE208EB0508}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SymbolCollector.Android.Library", "src\SymbolCollector.Android.Library\SymbolCollector.Android.Library.csproj", "{66444A27-B1C2-479F-9BE5-9B74113C1BE6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -127,5 +129,17 @@ Global
 		{48BB7372-FB0A-46DE-B66C-FDE208EB0508}.Release|x64.Build.0 = Release|Any CPU
 		{48BB7372-FB0A-46DE-B66C-FDE208EB0508}.Release|x86.ActiveCfg = Release|Any CPU
 		{48BB7372-FB0A-46DE-B66C-FDE208EB0508}.Release|x86.Build.0 = Release|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Debug|x64.Build.0 = Debug|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Debug|x86.Build.0 = Debug|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Release|x64.ActiveCfg = Release|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Release|x64.Build.0 = Release|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Release|x86.ActiveCfg = Release|Any CPU
+		{66444A27-B1C2-479F-9BE5-9B74113C1BE6}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
