+++ src/SymbolCollector.Android/MainActivity.cs
@@ -1,12 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Android.App;
 using Android.Content.PM;
 using Android.OS;
-using Android.Runtime;
 using Android.Systems;
 using Android.Views;
 using Android.Views.InputMethods;
@@ -14,14 +12,13 @@ using Android.Widget;
 using AndroidX.AppCompat.App;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Http;
 using Sentry;
-using Sentry.Extensibility;
-using Sentry.Protocol;
 using SymbolCollector.Core;
+using SymbolCollector.Android.Library;
 using AlertDialog = Android.App.AlertDialog;
 using OperationCanceledException = System.OperationCanceledException;
 using Xamarin.Essentials;
+using Host = SymbolCollector.Android.Library.Host;
 
 namespace SymbolCollector.Android
 {
@@ -29,7 +26,7 @@ namespace SymbolCollector.Android
         ScreenOrientation = ScreenOrientation.Portrait)]
     public class MainActivity : AppCompatActivity
     {
-        private string _friendlyName;
+        private readonly string _friendlyName;
         private readonly IHost _host;
         private readonly IServiceProvider _serviceProvider;
         private readonly ITransaction _startupTransaction;
@@ -251,95 +248,30 @@ namespace SymbolCollector.Android
 #pragma warning disable 618
             _friendlyName = $"Android:{Build.Manufacturer}-{Build.CpuAbi}-{Build.Model}";
 #pragma warning restore 618
-            StructUtsname? uname = null;
-
-            SentryXamarin.Init(o =>
-            {
-                o.TracesSampleRate = 1.0;
-                o.MaxBreadcrumbs = 200;
-                o.Debug = true;
-                o.DiagnosticLevel = SentryLevel.Debug;
-                o.AttachStacktrace = true;
-#if DEBUG
-                // It's 'production' by default otherwise
-                o.Environment = "development";
-#endif
-                o.Dsn = "https://2262a4fa0a6d409c848908ec90c3c5b4@sentry.io/1886021";
-                o.SendDefaultPii = true;
-
-                // TODO: This needs to be built-in
-                o.BeforeSend += @event =>
-                {
-                    const string traceIdKey = "TraceIdentifier";
-                    switch (@event.Exception)
-                    {
-                        case var e when e is OperationCanceledException:
-                            return null;
-                        case var e when e?.Data.Contains(traceIdKey) == true:
-                            @event.SetTag(traceIdKey, e.Data[traceIdKey]?.ToString() ?? "unknown");
-                            break;
-                    }
-
-                    return @event;
-                };
-            });
+            _host = Host.Init();
+            _serviceProvider = _host.Services;
 
             var tran = SentrySdk.StartTransaction("AppStart", "activity.load");
             _startupTransaction = tran;
 
-            // TODO: This should be part of a package: Sentry.Xamarin.Android
+            StructUtsname? uname = null;
+            try
+            {
+                uname = Os.Uname();
+                _friendlyName += $"-kernel-{uname?.Release ?? "??"}";
+            }
+            catch (Exception e)
+            {
+                SentrySdk.AddBreadcrumb("Couldn't run uname", category: "exec",
+                    data: new Dictionary<string, string> {{"exception", e.Message}}, level: BreadcrumbLevel.Error);
+                // android.runtime.JavaProxyThrowable: System.NotSupportedException: Could not activate JNI Handle 0x7ed00025 (key_handle 0x4192edf8) of Java type 'md5eb7159ad9d3514ee216d1abd14b6d16a/MainActivity' as managed type 'SymbolCollector.Android.MainActivity'. --->
+                // Java.Lang.NoClassDefFoundError: android/system/Os ---> Java.Lang.ClassNotFoundException: Didn't find class "android.system.Os" on path: DexPathList[[zip file "/data/app/SymbolCollector.Android.SymbolCollector.Android-1.apk"],nativeLibraryDirectories=[/data/app-lib/SymbolCollector.Android.SymbolCollector.Android-1, /vendor/lib, /system/lib]]
+            }
+
             SentrySdk.ConfigureScope(s =>
             {
                 s.Transaction = tran;
-                s.User.Id = Build.Id;
-#pragma warning disable 618
-                s.Contexts.Device.Architecture = Build.CpuAbi;
-#pragma warning restore 618
-                s.Contexts.Device.Brand = Build.Brand;
-                s.Contexts.Device.Manufacturer = Build.Manufacturer;
-                s.Contexts.Device.Model = Build.Model;
-
-                s.Contexts.OperatingSystem.KernelVersion = uname?.Release;
-
-                s.SetTag("API", ((int)Build.VERSION.SdkInt).ToString());
-                s.SetTag("app", "SymbolCollector.Android");
-                s.SetTag("host", Build.Host ?? "?");
-                s.SetTag("device", Build.Device ?? "?");
-                s.SetTag("product", Build.Product ?? "?");
-#pragma warning disable 618
-                s.SetTag("cpu-abi", Build.CpuAbi ?? "?");
-#pragma warning restore 618
-                s.SetTag("fingerprint", Build.Fingerprint ?? "?");
-
-#pragma warning disable 618
-                if (!string.IsNullOrEmpty(Build.CpuAbi2))
-#pragma warning restore 618
-                {
-#pragma warning disable 618
-                    s.SetTag("cpu-abi2", Build.CpuAbi2 ?? "?");
-#pragma warning restore 618
-                }
-#pragma warning restore 618
-
-#if DEBUG
-                s.SetTag("build-type", "debug");
-#elif RELEASE
-                s.SetTag("build-type", "release");
-#else
-                s.SetTag("build-type", "other");
-#endif
-                try
-                {
-                    uname = Os.Uname();
-                    _friendlyName += $"-kernel-{uname?.Release ?? "??"}";
-                }
-                catch (Exception e)
-                {
-                    SentrySdk.AddBreadcrumb("Couldn't run uname", category: "exec",
-                        data: new Dictionary<string, string> {{"exception", e.Message}}, level: BreadcrumbLevel.Error);
-                    // android.runtime.JavaProxyThrowable: System.NotSupportedException: Could not activate JNI Handle 0x7ed00025 (key_handle 0x4192edf8) of Java type 'md5eb7159ad9d3514ee216d1abd14b6d16a/MainActivity' as managed type 'SymbolCollector.Android.MainActivity'. --->
-                    // Java.Lang.NoClassDefFoundError: android/system/Os ---> Java.Lang.ClassNotFoundException: Didn't find class "android.system.Os" on path: DexPathList[[zip file "/data/app/SymbolCollector.Android.SymbolCollector.Android-1.apk"],nativeLibraryDirectories=[/data/app-lib/SymbolCollector.Android.SymbolCollector.Android-1, /vendor/lib, /system/lib]]
-                }
+                s.SetTag("friendly-name", _friendlyName);
 
                 if (uname is { })
                 {
@@ -351,53 +283,17 @@ namespace SymbolCollector.Android
                         uname.Sysname,
                         uname.Version
                     };
+                    s.Contexts.OperatingSystem.KernelVersion = uname.Release;
                 }
-            });
-
-            // Don't let logging scopes drop records TODO: review this API
-            HubAdapter.Instance.LockScope();
-
-            // TODO: doesn't the AppDomain hook is invoked in all cases?
-            AndroidEnvironment.UnhandledExceptionRaiser += (s, e) =>
-            {
-                e.Exception.Data[Mechanism.HandledKey] = e.Handled;
-                e.Exception.Data[Mechanism.MechanismKey] = "UnhandledExceptionRaiser";
-                SentrySdk.CaptureException(e.Exception);
-                if (!e.Handled)
-                {
-                    SentrySdk.Close();
-                }
-            };
-
-            var iocSpan = tran.StartChild("container.init", "Initializing the IoC container");
-            var userAgent = "Android/" + GetType().Assembly.GetName().Version;
-            _host = Startup.Init(c =>
-            {
-                // Can be removed once addressed: https://github.com/getsentry/sentry-dotnet/issues/824
-                c.AddSingleton<IHttpMessageHandlerBuilderFilter, SentryHttpMessageHandlerBuilderFilter>();
-
-                c.AddSingleton<AndroidUploader>();
-                c.AddOptions().Configure<SymbolClientOptions>(o =>
-                {
-                    o.UserAgent = userAgent;
-                    o.BlackListedPaths.Add("/system/build.prop");
-                    o.BlackListedPaths.Add("/system/vendor/bin/netstat");
-                    o.BlackListedPaths.Add("/system/vendor/bin/swapoff");
-                });
-                c.AddOptions().Configure<ObjectFileParserOptions>(o =>
-                {
-                    o.IncludeHash = true; // Backing store sorted format does not support hash distinction yet.
-                    o.UseFallbackObjectFileParser = false; // Android only, use only ELF parser.
-                });
-            });
-            iocSpan.Finish();
-            _serviceProvider = _host.Services;
-
-            SentrySdk.ConfigureScope(s =>
-            {
-                s.SetTag("user-agent", userAgent);
-                s.SetTag("friendly-name", _friendlyName);
-                s.AddAttachment(new ScreenshotAttachment());
+#if DEBUG
+                s.SetTag("build-type", "debug");
+                // It's 'production' by default otherwise
+                s.Environment = "development";
+#elif RELEASE
+                s.SetTag("build-type", "release");
+#else
+                s.SetTag("build-type", "other");
+#endif
             });
         }
 
@@ -405,36 +301,6 @@ namespace SymbolCollector.Android
         {
             base.Dispose(disposing);
             _host.Dispose();
-        }
-
-        private class ScreenshotAttachment : Attachment
-        {
-            public ScreenshotAttachment()
-                : this(
-                AttachmentType.Default,
-                new ScreenshotAttachmentContent(),
-                "screenshot",
-                 "image/png")
-            {
-            }
-
-            private ScreenshotAttachment(
-                AttachmentType type,
-                IAttachmentContent content,
-                string fileName,
-                string? contentType)
-                : base(type, content, fileName, contentType)
-            {
-            }
-
-            private class ScreenshotAttachmentContent : IAttachmentContent
-            {
-                public Stream GetStream()
-                {
-                    var screenshot = Screenshot.CaptureAsync().GetAwaiter().GetResult();
-                    return screenshot.OpenReadAsync().GetAwaiter().GetResult();
-                }
-            }
         }
     }
 }
