+++ src/SymbolCollector.Android/Properties/AndroidManifest.xml
